feat(shell.nix/ci_check): move the cargo checks into nix scripts.
This is the first step towards a real test runner instead of an
ad-hoc shell script to do ci checks.

The execline stuff is used to make it easier to generate scripts;
since it accepts a nix-style list, it’s a lot easier to generate than
string-based shell scripts (without fickly string interpolation bugs).