feat(shell.nix/ci_check): move carnix test into nix/ci and delete
Every single test was migrated over to `nix/ci`, so we can remove the
`ci_check` bash function. Instead, use the `ci` alias or execute the
nix-build of `ci.testsuite` to run the tests.

We remove the nix nixpkgs-fmt script because it doesn’t play well with
execline scripts, it reformats the lists too naively into one item per
line.

A little bit of setup is added to be able to do checks that don’t
require runtime dependencies at build-time (so they can be cached).