fix(tests/integration/direnv): add "outputs" and "system"
Nix derivations also contain these two fields, which are set as
environment variables, too.
This only appears inside of the ci testsuite with a fairly empty
environment, and it’s kind of a mystery why the test only fails there.

Restructures the test a bit to keep the values for assert outputs.