feat(watch): remove irrelevant distinction of no event paths
We only returned a `Result` from `watcher::process()` because we
wanted to distinguish between no new paths in the Event, but since we
are already returning `None` when there was no path in
the *interesting* paths, we don’t really need that.

This has some fortunate consequences, namely it makes it possible to
remove `EventError` alltogether, which propagates further into
build_loop, makes the `translate_reason` wart go away, and then even
makes it possible to remove the unknown error from both the enum and
the varlink interface.

This is a breaking change to the varlink, but I’m not on board
guaranteeing the stability of this interface, so frankly I don’t care.