chore(daemon): remove From instance for LoopHandlerEvent
I’m not a fan of From instances. Explicit is better than implicit.

This makes it kind of obvious that we shouldn’t be sending
`LoopHandlerEvent` on the channel in the first place, because we only
ever send the one kind (`BuildEvent`) from our build loop anyway.
However, it’s not as easy to map things over a channel after the fact,
it requires yet another thread, and threads are system threads so not
for free.