chore(build_loop): remove mystical output gc dir check
This check for whether the gc output path exists has been there in
forever, undocumented. Then I added a TODO asking for clarification
another eternity ago, but even staring at it for extended amounts of
time does not give me any clues as to why it is there.

I say in face of the obvious simplification of the code we can remove
it for now, the only effect it could have is rebuilding more often.
But since we are *already* rebuilding way too often (until we improve
the build loop which I’m in the process of doing), we can risk that.
Simple code is more important for the restructure of the build loop.