chore(Cargo.toml): bump dependencies & trim closure
It’s time for an update run!

I defined a minimum supported rust compiler that we want to
support (1.41.0) and then went through all of our dependencies to
update to the according version and make sure unnecessary features are
disabled.

The list also got sorted by core features vs nice-to-have, so it’s
easier to see which ones are more important to update correctly.

Stripping features from e.g. doctest (which we only use for one tiny
test) made our closure shrink a bit! In general it’s worth
investigating this stuff. I also moved it to dev-dependencies, now
cargo won’t build it if you don’t run the tests.