fix(build_loop): always schedule at most one other build
This switches the build loop from blockingly running a build for every
event that ever appeared, to always at most running one additional
build as soon as the current one finishes.

Whenver we start a new build, we take the *current* state of the
environment to run nix, which means that all events that happen
while a build is running (e.g. pings, filesystem events &c.) should be
collapsed into *one* (1) extra build to be scheduled right after the
current build finishes.

So there is now a stateful `current_build` variable, which contains
one `BuildState` that is mutably updated from `NotRunning` to
`Running` and if it’s already running to `RunningAndScheduled`. Then
the “build finished” event becomes just another handler in our giant
`select!`, meaning that as soon as a build finishes we can either go
back to `NotRunning` or go to `Running` if there had been another
build scheduled.

For now this means inlining `once_with_send` into the build loop,
further refactors to follow to make the logic easier to understand.