fix(build_loop): remove initial unconditional build
The build loop would previously drain pings in order to prevent
building the project twice immediately. That is well-meaning, but just
a workaround for the fact, that we would already build once.

However, we can remove the unconditional first build, since external
processes would ping us anyway. That is, the build loop would not even
be started unless some external process has pinged the daemon.

This leads to `lorri watch` not building initially, but we can just do
the same as external clients would do, and ping the build loop.