feat(build_loop,ping): add a rebuild instruction to the ping
When we ping the build_loop, we so far left it unspecified whether
`IndicateActivity` will always trigger a build or only in certain
situations.

After the refactoring of build_loop, the observable behaviour of lorri
changed somewhat, in that every ping will now schedule a rebuild,
which for direnv users is on every `lorri direnv` call, e.g. on
switching buffers or executing a command.

This commit introduces a flag to IndicateActivity which makes it clear
to the builder whether it should rebuild unconditionally (`lorri
ping`) or only if there is no build running at the moment.

It does not change the behaviour yet, everything is called with
`Rebuild::Always`.