fix(ops/direnv): only trigger rebuild if the project was not watched
After fixing up the build loop people suddenly started noticing that
lorri was evaluating every time something ran `lorri direnv`, which
could potentially be every time the user switched between buffers in
the editor.

This is not the intended behaviour, since we should run an
unconditional build only the first time the project is added to the
watcher, and after rely on the watcher to notify us of any file
changes (or the user running `lorri internal ping` to force a
rebuild).

This removes the `Rebuild::OnlyIfNotBuilding` that we added in the
previous build, since it was not the required semantics, and replaces
it by `Rebuild::OnlyIfNotYetWatching`, which determines which build
loops to ping from the daemon.

A small refactor removes the `Handler` struct, since it would only
carry the `tx_ping` which is only populated and used in one place
anyway.

As a nice side-effect, the logic of how and when to add a new
build_loop and how and when to ping moves into a pretty concise match
on a tuple, which should be easier to understand than
`entry(…).or_insert_with(…)`.