fix(thread): don’t panic if the death message can’t be sent
If the other side of that channel is closed, it means that the pool
propagated a panic, so we should just log the fact instead of panicing
again.

The logging then needs to happen via an explicit logger, since
slog_scope doesn’t seem to reach into the thread. We add explicit
loggers to both `Pool` and `Async`, and adjust calls. In order to call
these in tests, we provide a simple stderr logger setup via
`logging::test_logger()`.