feat(socket): Add a native daemon/client protocol
tl;dr: varlink bad, serde good, use `LORRI_BACKEND=native` to make
`lorri daemon` and `lorri internal ping` use the serde-based protocol.
All the other commands to follow.

~~~

This re-adds a slightly modified version of the native communication
protocol between lorri server and client(s) that was removed in
b0cbe27c4b762bef8d820cd146da75392e1880c5 in favor of a varlink-based
RPC format.

I was skeptical back when it was introduced, since the arguments in
favor of varlink where mostly “we don’t have to maintain any custom
socket code” and “RPC is better than whatever the current socket
protocol does”, without actually trying to work with the existing
socket code first.

As it turns out, varlink has multiple weaknesss that make working with
it a pain in the ass:

1. It is JSON-based.

Thus is cannot encode arbitrary bytes like filepaths. lorri is full of
filepaths. Every event that happens contains filepaths. Also program
output. Which is also random bytes.

Which lead to hundreds of `TryFrom` instances which convert paths to
utf-8, crashing lorri in case they are not utf-8 clean. Also all these
conversions from and to strings were written by hand.

There is a format that can encode arbitrary rust structs and enums,
and it’s used by most rust code that does any serialization, it’s
called serde. It has a binary representation and also a json one if
you really need the debuggability. Which brings us to

2. Varlink is a format made for public interfaces

It is based on the assumption that there is external consumers of the
API which want to auto-generate some code to talk to it. But the lorri
server/client protocol is not that. It’s internal. We don’t want to
talk to other tools through this API, we want to talk to a lorri
client/daemon of exactly the same version.

Yes, we have `lorri internal stream-events`, which is a prototype of
an external interface to the build events, and which currently uses
varlink, but that is entirely an implementation detail. In fact, how
the client (`stream-events`) talks to the server (`lorri daemon`) and
how it prints those messages to the user should probably be different
data types in the first place anyway, so that we can maintain
backwards-compat while being able to refactor internally.

3. Varlink does not understand ADTs and rust enums

Json does not do enums. There is a bunch of encodings for them, and
they can also be done manually, but varlink is a “protocol” written by
go programmers, who do not know how ADTs work generally. Thus it does
not have any good enum type. Thus we need to manually encode/decode
enums into the format it understands. There is a rust library which
does this automatically, serde.

4. Varlink has bad documentation, no specification, and little users

Initially when we introduced it, it might have looked as if varlink
was picking up steam, but the simple fact is that varlink
documentation is very sparse, and hasn’t been improved since we first
introduced it in 2019. Probably because there is a bunch of formats
with better adoption out there that actually can encode public APIs
and have tooling to support things like versioning, like protobuf or
thrift.

Also the part that varlink uses a custom format, and the EBNF is the
best you get, no examples of usage, no patterns, no FAQ.

Q: So why not switch to protobuf?

A: It was an idea, and may yet be a good idea, but only for the public
interface of lorri. For the internal interface where we essentially
have a unix socket that `accept()`s in a loop, and can require users
to keep the version of the server and client in exact sync, there is
little reason to introduce manual encoding and parsing steps that come
with the benefit of backward/forward compat.

Q: Why is this so much code?

A: The code, as the one that was originally replaced, has a module
that describes a generic typesafe
reader/writer (`src/socket/read_writer.rs`), and a higher-level module
with a server and a client implementation that uses
it (`src/socket/communicate.rs`).

The read/writer module is handling the type-based serialization and
reading/writing typed messages to a socket via the serde bincode
format. It supports timeouts, which complicates the code slighly but
is reqiured for not blocking indefinitly. The code should be
relatively easy to understand.

The communicate module implements the meat of the communication, and
it also defines all different communications that can happen between
the various clients and the server, using the underlying read/writer.
There is a bit of trait wrangling to make sure that a communication
always binds the same two request/response types together, and a bit
of code to define a typesafe listener and clients for all handlers.

The most messy module right now is `src/socket.rs`, which implements
the actual accept loop, and tries its best to not run out of threads
while doing that. lorri might run for a long time and be called many
times (e.g. via `lorri direnv`), so we don’t want to generate too many
thread zombies. We could opt for a thread pool instead, which doesn’t
have this problem.

Q: This is just the ping event, right?

Yes, this commit only goes to the point of where the code was when it
was first replaced with varlink, which is the ping event. The rest of
the events is going to be implemented in (a) following commit(s).

Let’s look forward to dropping varlink in a bit.