feat(ops): add native backend support for stream-events
Since we are using serde, we need to add a few more deserialize calls
to types, but it comes for free.

There was a particularly abhorrent Serialize instance for a path type,
which is removed.

I’m so looking forward to removing all the horrible TryFrom instances
and other war crimes.