chore(nix): patch carnix to not explode on unknown fields
We should get away from carnix as fast as possible, it hasn’t been
maintained and the quality of the tool is questionable. I wasn’t even
able to reach nest.pijul.org, and this is the only place the source is
hosted.

The error this works around was undebuggable from the message itself,
and I’m still not entirely sure what exactly happens, since there is
no source code comments anywhere to be found. Very bad software.