chore(/): remove any trace of varlink
Finally. See previous commits for reasoning. Nothing in the public
interface should change.

The client code for the native protocol still needs to be improved
somewhat so that it does not die on `unwrap()` or similar.

This removes some tests that were just hopelessly broken, e.g. the
“integration” “test” that had to poll for the socket to be available
after starting the server, instead of propagating out readiness. Not
even talking about the fact that these tests should just call the
actual binaries.