feat(ops/stream-events): split off the json Serialize instances
Previously we’ve been using the same Serialize instance for the
internal “over the wire” protocol between daemon/client and for the
stream-events command json serialization.

This needs to be split up if we ever want to stabilize the interface.

Thus a first step, which keeps the inner structure of `Event`, but
stubs out all the inner types that have to be converted for the json
interface.

This pattern looks quite nice, even though implementing `map()` for
each of these interface types is a bit of boilerplate.