fix(ops): remove all `unwrap()`s from ops, add anyhow
After getting rid of varlink the new clients were temporarily created
with lots of `unwrap()`ping, but we can do better.

In the spirit of https://github.com/nix-community/lorri/issues/12, I
first tried to add the `failure` crate, but after a while I noticed
that the usability is really low. Even though it has documentation,
the documentation doesn’t talk about how to make backtraces work.
Additionally, you would have to build your own error pretty printing.

Based on the recommendation in the issue, I switched the error
definitions to `anyhow` instead, which feels a lot more modern and
*does* integrate better with the `std::error::Error` trait (even
though backtraces don’t work on stable, but they didn’t for `failure`
either, and we will get them for free in time).

There is a small trait `ExitAs`, that gives an automatic `From`
instance to every error type that implements it and `Error`, which is
nice. Though I couldn’t get it to work with the `Context` trait like I
did with the `Context` type from `failure.

~~~

A small improvement was done to the `thread::Pool`, in addition to
only returning instantly when a thread panics, we now also return when
any thread returns an `Err`. This allows us to actually transmit typed
errors, and use the `?` operator directly in the thread closure.

~~~

In ops itself, we factor out all the client connections into the
`daemon::client` module, and replace all unwraps by ExitErrors.
This gives us back the nice exit errors we had before, plus error
traces for anything that implements `Error` and uses `#[source]`.
Nice.

The `lorri direnv` call was still broken when the daemon was not
running, this fixes that as well.