chore(/): 1.3.1 -> 1.4.0 & add changelog entry
It’s a major version bump because we remove a CLI flag (even though it
was probably not used much).

The internal change might also expose new bugs, so let’s keep it safe
and tell people that the behaviour might have changed.
Plus `internal stream-events` also has seen some damage, so it would
be unfair to make this a patch release.