chore(/): fix clippy warnings
It turns out that, in fact, we aren’t always using a recent clippy,
but there were real clippy warnings that we needed to fix.

The `crossbeam_channel` problem would be fixed in a newer version, but
the version used by `notify` is older, and updating `notify` to a
newer pre-release won’t be a trivial fix, since they changed how
debouncing works and it might break our notification code (at least it
requires a bunch of testing).