chore(/): fix clippy warnings that only appear on CI
I am going crazy. There is literally *no* divergence between the local
testsuite run and the one on CI, the whole environment is empty, the
versions are all exactly the same, using the same pinned nixpkgs, and
to top it off everything is run with an empty environment.

Yet, clippy shows these warnings on CI, and not when running it
locally. I want to curse whoever is responsible.