feat(run_async): start Asyncs that linger after they drop
In most cases, blocking on the thread join in the `Drop` instance is
the right thing to do, to make sure the inner task finishes before
anything else is done.

In some cases however, we are just spawning the async as an
intermediate pipline step, and are not interested in it when it goes
out of scope, for example when the program finishes. Then we can
`run_and_linger`.