refact(daemon,socket}: move thread spawning out of `accept()`.
Best viewed with whitespace ignored, since I had to indent a big chunk
of handlers.

`accept()` spawning a thread itself was very awkward, since it breaks
out of the standard “block, and if the user wants to run it in
parallel, they have to wrap it in an async” pattern.

This way we can wrap the handlers in an async instead of a thread in
the next step.