(feature) do not watch recursively directories used through builtins.readDir
Currently if a nix file depends on `builtins.readDir path`, `path` is
recursively watched. This is wasteful because if "path/foo/bar" is modified,
`builtins.readDir path` does not change. We should reevaluate
only if the list of direct children of `path` changed.

This patch separates paths to be watched in paths to be watched
recursively and patches to be watched non recursively. The latter is
used for builtins.readDir.