(feature) add support for builtins.filterSource
nix does not log anything when evaluating builtins.filterSource,
contrary to normal paths. We have to wrap builtins.filterSource with
builtins.trace like builtins.readFile.

The easy solution would be to trace "lorri read: $path", but then we
also watch filtered out files. Reimplementing the full logic of
filterSource in nix would be cumbersome, but we can find a treadeoff.
This implementation is designed to ignore .git when using nix-gitignore:
the top-level directory is watched as if it was readDir'ed, and only the
children which are kept by filterSource (ie, not .git) are watched
recursively. This way you can run git fetch in the filterSource'd
directory and not reevaluate. I suppose nix-gitignore accounts for a
large part of filterSource usage, as filterSource is kind of painful
to use.