fix(daemon/server): break out of the stream-events handler
If the client disengages, we need to stop the handler immediately,
otherwise we keep it running in the for-loop for all times.

This fixes a problem where any call of the
`lorri internal stream-events --kind snapshot` client
would drop the connection, and the handler would trace it but not stop
itself.

Fixes https://github.com/nix-community/lorri/issues/35