fix(/): reenable logging in many contexts
Gets rid of the `slog_scope` crate, which silently dropped log
messages in cases where the global logger wasn’t fully initialized (a
lot of cases as it turns out).

Instead we pass the logger through manually at the appropriate places.
This is a bit of work, but keeps us honest about which code is
logging, and will always work trivially.

In cases where most functions already took `self` I opted for cloning
the logger into the class struct, but in general it’s probably better
to manually pass it through as a function argument.