chore(nix): update minimum tested nixpkgs version 19.09 -> 20.09
This will bump the stable rust version to 1.45.2, which is the lowest
version we are gonna test with.

With some effort 19.09 could probably also be possible, and in general
I want to keep it compatible as far back as possible. But too many
transitive dependencies have advanced by now, using newer compiler
features. :(

vec1 is the only package that already broke compat with 1.45, so we
restrict its version.