fix(project): move user root dir check to toplevel
Instead of only checking whether we can create the user gcroot when we
actually want to create a root in it, we check once in the beginning
of setting up lorri and then pass the path through, assuming it
exists.

This improves the error message when the necessary setup is not yet
there (it’s an enironment problem which can be fixed on the first
start of lorri).