fix(.github): separate CI run from build
This way it’s clearer which step fails, and it’s faster to scroll to
the test output. We could(?) make every test into its own step at one
point, that would be good I think.

Given that we use nix for both, we could import the tests and parse
their attributes from the list to generate our CI run.