refact(*): Remove ContentAddressable from Project & pass manually
We only pass the `cas` through the `Project` constructor, never
actually using it in the project `impl`!. This means we don’t need to
burden `Project` with the cas, since it’s only used by the builder in
the end. Insead, put it on the build loop for now.

Funnily enough, `StartUserShell` does not need the project init
anymore, cause it only wanted the `cas` from it.