feat(ops/info): add a daemon status
lorri info will try to connect to the daemon, and print a status
message with either “daemon is running” or an error message if it
cannot connect to the socket.

We introduce a new communication message type, `DaemonInfo`, which is
used for that connection. Currently it is empty, but we can add more
info to show easily (e.g. the build status).