feat(ops/shell): print a nice warning for unsupported shells
We’ve only really used `shell` for the first few months, then expected
people to switch to `direnv`. So we only have real support for `bash`
and `zsh`. We should note that.

By not shadowing `project` in the main function, I uncovered that we
missed out on using the nested logger for `Info`.