refact(tests/gc): repair the gc integration test
What an amazing piece of code.

Apparently it wasn’t even worth the work to remove all `dbg!` macros
before submitting it upstream.

Get rid of adding `stderr` to the checked output.

I have a very trigger-happy finger that was very close to simply
deleting this whole file without a second thought, but I suppose
having some kind of integration test for this is at least a little
useful.