fix(project): create project for each gc root
Instead of manually listing the files and doing some ad-hoc reading of
files to list every gc dir, use our existing project setup to fetch
this information.

This switches around the logic a bit, because I couldn’t be bothered
to keep it all the same. Especially around the liveness of nix files.
Basically there might be a nix_file symlink file, but there might also
not. And if it’s there, there might be a nix_file behind it or not.
Same for the timestamp and the gc root symlink.

For flakes this made clear that we have to save the installable with
the gc root, otherwise we cannot say after the first invocation which
part of the flake was evaluated. It’s gonna be interesting because
that means you can have different caches per project. Which one do you
choose? Which one is evaluated by `lorri direnv`?