refact(build_loop): rewrite build loop logic
Instead of the pretty complicated `BuildState` enum stuff (which did
*not* gel very well with mutability), we switch over to two mutable
`Option<>` variables which we drain and set.

This means there’s one extra state that should not happen (not running
& scheduled), but the logic is now a lot easier to follow.

We also pull the handling logic out of the `select!` macro, because in
the next step we are gonna switch to `tokio::sync::mpsc` which
requires a `&mut Receiver` for `recv()`.