refact(*): (partial) rewrite to async with tokio
Rewrites everything except the watcher and the build code to use
async.

The internal client/server communication protocol is switched back to
use plain line-based json instead of bincode, to make it easier to
debug with `strace` (and because integrating bincode into async was
harder than necessary).

The biggest non-trivial change is in the `build-loop` module.
Everything else should be mostly straight-forward based on compiler
errors.

Any time we `select!` on a channel from sync code, we use `crossbeam`,
any time we `select!` from async code, we use `tokio::sync`.

We might want to throw out the last use of our own `Async` module at
one point and replace it with a simple direct OS-thread.

There could be some cases of blocking code that were overlooked, so
some more in-depth testing will be needed.