feat(main): improve crash error message
If a panic happens, we now directly present a clickable link that
opens an issue on Github, containing the error in the URL if it fits,
or at least the title (and prompting the user to copy the error
message themselves).

This means we can get rid of the `human-panic` vendor, since I inlined
the interesting bits (which are not all that many tbh, mainly the
backtrace logic).