fix(communicate,read_writer): preserve any BufReader state
Our `read_writer` and `communicate` stuff was buggy, in that we always
constructed an new `BufReader` to read line-by-line from the unix
socket, and then extracted the raw socket again after each read.

This means if we read over a line boundary we’d silently drop every
byte after the first line.

We never ran into this before, because we always interleaved a read
with a write, but since I removed the `write` from the snapshot
reader, there would be two `read`s right after each other, and in some
cases (depending on OS scheduling) that would lead to running into the
bug above.

Now we keep the state in `ReadWriterState` and make sure we never
toss it until shutdown.