refact(*): Use tokio current thread runtime everywhere
(diff best viewed without whitespace changes)

lorri does not really need the multi-threaded runtime.
Instead, we can move everything to use `spawn_local`, meaning we don’t
need a `Send` constraint on our futures, because no value will be
transferred between threads, major profit.

Annoyingly, the `tokio::test` macro does not support putting things on
a `LocalSet` or `LocalRuntime` for us, so we have to wrap everything
in our setup manually. But that’s fine.

I rewrite the `.retain` logic to use optional channels that get
set to None and cleaned of those afterwards, so we can .await on the
channel send.