feat(src/project): add test for creation of indirect gc roots
This tests the creation of gc roots pretty directly.

Only problem is that `nix-store --delete` needs to list
`/nix/store/.links` every time, this is 1.5 million files on my
system, which takes about 5 seconds, and the test has to run it twice
ugh. Not sure what to do about that.