feat(sqlite): test the migration to sqlite
Adds a test that assures the migration code works according to the
existing gc dirs.

Now, this does construct a separate gc dir, but since we are gonna
move the `project` module away from doing things in the file system as
much, we will have to manually construct a pre-migration dir anyway
for the migration test.

This should test all stored bits from the gc dirs, and compare the
migration of both flakes and shell.nix files.