feat(sqlite): finish migrating to sqlite
This finishes the initial migration to sqlite.

1) We add a check whether we are already migrated.
2) If we successully manage to migrate, we delete the nix file
backlinks to clean up our cache directories.

The migration test now checks that the backlinks are sucessfully
removed as well.

The database file is never cleaned, and renamed to `lorri.sqlite`.

From this change on, we’ll have to think about and test migration
paths between database versions, somehow …