doc(lorri.scd): lorri.1 -> lorri.scd, drop self-upgrade section
The manpage was previously written in the `.mdoc` format, which, while
“semantic” at its core, is completely inscrutable to the average
person. It was the best tool available at the time.

However, in the meantime Drew DeVault create `scdoc`
(see https://drewdevault.com/2018/05/13/scdoc.html)
which has most interesting features and is 1/10th of the complexity.

So let’s migrate our manpage.

I did not migrate the `self-upgrade` section, because we removed that
subcommand in a previous commit.