feat(ops): set up `lorri internal varlink stdin`
This subcommand will implement a varlink service.

Right now it can be communicated with over stdin/out, but the library
provides an easy way of registering a unix socket and systemd socket
activation works as well.

The varlink file is code-generated and then called from our source
code, making sure the types are as declared in the interface file.